Add option to "naturalize" ThingTalk in MultiWOZ
Replaces ThingTalk tokens that are not used in natural langauge with
shorter and more natural tokens. This transformation is applied to both
context (input) and answer (output). The output of the model is changed
back to the original format
so that this transformation remains transparent to
the outside evaluation scripts.